Consolidating rabbit modules
