+++ src/app/beer_garden/instances.py
@@ -11,7 +11,7 @@ import beer_garden
 import beer_garden.db.api as db
 import beer_garden.queue.api as queue
 from beer_garden.events import publish_event
-from beer_garden.queue.rabbitmq import get_routing_key, get_routing_keys
+from beer_garden.queue.rabbit import get_routing_key, get_routing_keys
 
 logger = logging.getLogger(__name__)
