+++ src/app/beer_garden/queues.py
@@ -1,14 +1,12 @@
 # -*- coding: utf-8 -*-
 import logging
 
-from brewtils.errors import NotFoundError
 from brewtils.models import Events, Queue, System
-from pyrabbit2.http import HTTPError
 
 import beer_garden.db.api as db
 import beer_garden.queue.api as queue
 from beer_garden.events import publish_event
-from beer_garden.queue.rabbitmq import get_routing_key
+from beer_garden.queue.rabbit import get_routing_key
 
 logger = logging.getLogger(__name__)
